implemented switch to enable output deletion

+++ snap-performance-test/src/test/java/org/esa/snap/performance/util/PerformanceTestRunner.java
@@ -15,11 +15,13 @@ public class PerformanceTestRunner {
     List<PerformanceTest> tests;
     private List<PerformanceTestResult> results;
     private String outputDirectory;
+    private boolean deleteOutput;
 
-    public PerformanceTestRunner(List<PerformanceTest> tests, String outputDirectory) {
+    public PerformanceTestRunner(List<PerformanceTest> tests, String outputDirectory, boolean deleteOutput) {
         this.tests = tests;
         this.results = new ArrayList<>();
         this.outputDirectory = outputDirectory;
+        this.deleteOutput = deleteOutput;
     }
 
     public void runTests() {
@@ -36,11 +38,12 @@ public class PerformanceTestRunner {
                 logger.log(Level.SEVERE, "Test will be skipped: " + test.getClass().getSimpleName() + " - " + e.getMessage(), e);
             }
 
-            // TODO: implement switch as global config parameter to select if output should be deleted after every iteration
-            try {
-                TestUtils.deleteTestOutputs(this.outputDirectory);
-            } catch (IOException e) {
-                logger.log(Level.INFO, "Performance test output files could not be deleted: " + e.getMessage(), e);
+            if (deleteOutput) {
+                try {
+                    TestUtils.deleteTestOutputs(this.outputDirectory);
+                } catch (IOException e) {
+                    logger.log(Level.INFO, "Performance test output files could not be deleted: " + e.getMessage(), e);
+                }
             }
 
             System.gc();
